requirements.txt
coloredlogs==15.0.1
customtkinter==5.2.2
darkdetect==0.8.0
flatbuffers==25.12.19
humanfriendly==10.0
mpmath==1.3.0
numpy==2.2.6
onnxruntime==1.23.2
packaging==26.0
pillow==12.1.1
protobuf==6.33.5
sympy==1.14.0
tkinterdnd2==0.4.3
